Renamed logformatter to log_formatter
